reports/20260322-145232/results/key-vault/data-plane/python/pagination/azure-mcp/generated-code/requirements.txt
# Azure Key Vault Secrets SDK
# Documentation: https://learn.microsoft.com/python/api/overview/azure/keyvault-secrets-readme
azure-keyvault-secrets>=4.8.0

# Azure Identity for authentication
# Documentation: https://learn.microsoft.com/python/api/overview/azure/identity-readme
azure-identity>=1.15.0

# Optional: Azure Core for understanding ItemPaged internals
azure-core>=1.29.0
